requirements.txt
Flask
Flask-cors
Flask-SQLAlchemy
psycopg2
line-bot-sdk
requests
python-dotenv
